Make `CorsFilter` available to all `web` users
