+++ integration-tests/test-app/src/main/java/io/spine/web/test/given/CrossOriginResourceSharingFilter.java
@@ -20,78 +20,18 @@
 
 package io.spine.web.test.given;
 
-import com.google.common.net.HttpHeaders;
+import io.spine.web.CorsFilter;
 
-import javax.servlet.Filter;
-import javax.servlet.FilterChain;
-import javax.servlet.FilterConfig;
-import javax.servlet.ServletException;
-import javax.servlet.ServletRequest;
-import javax.servlet.ServletResponse;
 import javax.servlet.annotation.WebFilter;
-import javax.servlet.http.HttpServletResponse;
-import java.io.IOException;
 
-import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
+import static io.spine.web.CorsFilter.ANY_URL;
+import static io.spine.web.CorsFilter.NAME;
 
 /**
- * An servlet filter which appends the CORS headers to the {@code /command} and {@code /query}
- * HTTP responses.
+ * A filter which allows requests from any origin to interact with any endpoint of this server.
  *
- * <p>Requests from any origin are allowed. The requests are allowed to contain credentials and
- * the {@code Content-Type} header.
+ * @see CorsFilter
  */
-@WebFilter(filterName = CrossOriginResourceSharingFilter.NAME, urlPatterns = "*")
-public final class CrossOriginResourceSharingFilter implements Filter {
-
-    static final String NAME = "CORS filter";
-
-    @Override
-    public void doFilter(ServletRequest request,
-                         ServletResponse response,
-                         FilterChain chain) throws IOException, ServletException {
-        appendHeaders((HttpServletResponse) response);
-        chain.doFilter(request, response);
-    }
-
-    private static void appendHeaders(HttpServletResponse response) {
-        for (ResponseHeader header : ResponseHeader.values()) {
-            header.appendTo(response);
-        }
-    }
-
-    @Override
-    public void init(FilterConfig filterConfig) {
-        // NOP.
-    }
-
-    @Override
-    public void destroy() {
-        // NOP.
-    }
-
-    /**
-     * The HTTP headers which configure the cross-origin request handling.
-     */
-    private enum ResponseHeader {
-
-        ACCESS_CONTROL_ALLOW_ORIGIN(HttpHeaders.ACCESS_CONTROL_ALLOW_ORIGIN, "*"),
-        ACCESS_CONTROL_ALLOW_CREDENTIALS(HttpHeaders.ACCESS_CONTROL_ALLOW_CREDENTIALS, "true"),
-        ACCESS_CONTROL_ALLOW_HEADERS(HttpHeaders.ACCESS_CONTROL_ALLOW_HEADERS, CONTENT_TYPE);
-
-        private final String name;
-        private final String value;
-
-        ResponseHeader(String name, String value) {
-            this.name = name;
-            this.value = value;
-        }
-
-        /**
-         * Appends this header to the ginen {@link HttpServletResponse}.
-         */
-        private void appendTo(HttpServletResponse response) {
-            response.addHeader(name, value);
-        }
-    }
+@WebFilter(filterName = NAME, urlPatterns = ANY_URL)
+public final class CrossOriginResourceSharingFilter extends CorsFilter {
 }
